frontend/requirements.txt
streamlit==1.27.0
requests==2.31.0
pillow<10.0.0,>=7.1.0
python-dotenv==1.0.0
